refactor splom - add files for calc plot hover select helpers and update_scene
